BITSCOUT-27: initial qpid client support for fluentd
This is the first cut of the fluentd amqp input plugin.

There is one major problem: the proton ruby wrapper is calling C functions that
should block in poll() for efficiency, but is holding the GVL. To avoid holding
up the rest of fluentd I have set this to busy-loop rather than block but that
is not efficient.

I will rework the code to resolve this shortly. Other to-do items:

- ruby gem packaging
- better unit testing (currently starting the fluentd exe from the tests)